backend/requirements.txt
Django>=2.1.7,<2.2.0
djangorestframework>=3.9.1,<3.10.0
psycopg2>=2.7.7,<2.8.0
Pillow>=5.4.1,<5.5.0

flake8>=3.7.7,<3.8.0
requests
